refactor: move spellcheck widgets and manager to ui/spellcheck.py to remove Qt dependencies from utils
